move qnn loss to subfolder
