+++ src/squlearn/qnn/loss/qnn_loss_base.py
@@ -1,0 +1,362 @@
+"""Loss Base Classes for QNNs."""
+
+import abc
+from collections.abc import Callable
+from typing import Union
+import numpy as np
+
+
+class QNNLossBase(abc.ABC):
+    """Base class implementation for loss functions."""
+
+    def __init__(self):
+        self._opt_param_op = True
+
+    def set_opt_param_op(self, opt_param_op: bool = True):
+        """Sets the `opt_param_op` flag.
+
+        Args:
+            opt_param_op (bool): True, if operator has trainable parameters
+        """
+        self._opt_param_op = opt_param_op
+
+    @property
+    def loss_variance_available(self) -> bool:
+        """Returns True if the loss function has a variance function."""
+        return False
+
+    @property
+    @abc.abstractmethod
+    def loss_args_tuple(self) -> tuple:
+        """Returns evaluation tuple for loss calculation."""
+        raise NotImplementedError()
+
+    @property
+    def variance_args_tuple(self) -> tuple:
+        """Returns evaluation tuple for loss calculation."""
+        raise NotImplementedError()
+
+    @property
+    @abc.abstractmethod
+    def gradient_args_tuple(self) -> tuple:
+        """Returns evaluation tuple for loss gradient calculation."""
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def value(self, value_dict: dict, **kwargs) -> float:
+        """Calculates and returns the loss value."""
+        raise NotImplementedError()
+
+    def variance(self, value_dict: dict, **kwargs) -> float:
+        """Calculates and returns the variance of the loss value."""
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def gradient(
+        self, value_dict: dict, **kwargs
+    ) -> Union[np.ndarray, tuple[np.ndarray, np.ndarray]]:
+        """Calculates and returns the gradient of the loss."""
+        raise NotImplementedError()
+
+    def __add__(self, x):
+        """Adds two loss functions."""
+        if isinstance(x, QNNLossBase):
+            return _ComposedLoss(self, x, "+")
+        elif isinstance(x, float) or isinstance(x, int) or callable(x):
+            return _ComposedLoss(self, ConstantLoss(x), "+")
+        else:
+            raise ValueError("Only the addition with another loss functions are allowed!")
+
+    def __radd__(self, x):
+        """Adds two loss functions."""
+        if isinstance(x, QNNLossBase):
+            return _ComposedLoss(x, self, "+")
+        elif isinstance(x, float) or isinstance(x, int) or callable(x):
+            return _ComposedLoss(ConstantLoss(x), self, "+")
+        else:
+            raise ValueError("Only the addition with another loss functions are allowed!")
+
+    def __mul__(self, x):
+        """Multiplies two loss functions."""
+        if isinstance(x, QNNLossBase):
+            return _ComposedLoss(self, x, "*")
+        elif isinstance(x, float) or isinstance(x, int) or callable(x):
+            return _ComposedLoss(self, ConstantLoss(x), "*")
+        else:
+            raise ValueError("Only the addition with another loss functions are allowed!")
+
+    def __rmul__(self, x):
+        """Multiplies two loss functions."""
+        if isinstance(x, QNNLossBase):
+            return _ComposedLoss(x, self, "*")
+        elif isinstance(x, float) or isinstance(x, int) or callable(x):
+            return _ComposedLoss(ConstantLoss(x), self, "*")
+        else:
+            raise ValueError("Only the addition with another loss functions are allowed!")
+
+    def __sub__(self, x):
+        """Subtracts two loss functions."""
+        if isinstance(x, QNNLossBase):
+            return _ComposedLoss(self, x, "-")
+        elif isinstance(x, float) or isinstance(x, int) or callable(x):
+            return _ComposedLoss(self, ConstantLoss(x), "-")
+        else:
+            raise ValueError("Only the addition with another loss functions are allowed!")
+
+    def __rsub__(self, x):
+        """Subtracts two loss functions."""
+        if isinstance(x, QNNLossBase):
+            return _ComposedLoss(x, self, "-")
+        elif isinstance(x, float) or isinstance(x, int) or callable(x):
+            return _ComposedLoss(ConstantLoss(x), self, "-")
+        else:
+            raise ValueError("Only the addition with another loss functions are allowed!")
+
+    def __truediv__(self, x):
+        """Divides two loss functions."""
+        if isinstance(x, QNNLossBase):
+            return _ComposedLoss(self, x, "/")
+        elif isinstance(x, float) or isinstance(x, int) or callable(x):
+            return _ComposedLoss(self, ConstantLoss(x), "/")
+        else:
+            raise ValueError("Only the addition with another loss functions are allowed!")
+
+    def __rtruediv__(self, x):
+        """Divides two loss functions."""
+        if isinstance(x, QNNLossBase):
+            return _ComposedLoss(x, self, "/")
+        elif isinstance(x, float) or isinstance(x, int) or callable(x):
+            return _ComposedLoss(ConstantLoss(x), self, "/")
+        else:
+            raise ValueError("Only the addition with another loss functions are allowed!")
+
+
+class _ComposedLoss(QNNLossBase):
+    """Special class for composed loss functions
+
+    Class for addition, multiplication, subtraction, and division of loss functions.
+
+    Args:
+        l1 (QNNLossBase): First loss function
+        l2 (QNNLossBase): Second loss function
+        composition (str): Composition of the loss functions ("+", "-", "*", "/")
+
+    """
+
+    def __init__(self, l1: QNNLossBase, l2: QNNLossBase, composition: str = "+"):
+        super().__init__()
+        self._l1 = l1
+        self._l2 = l2
+        self._composition = composition
+        self._opt_param_op = self._l1._opt_param_op or self._l2._opt_param_op
+        self._l1.set_opt_param_op(self._opt_param_op)
+        self._l2.set_opt_param_op(self._opt_param_op)
+
+    def set_opt_param_op(self, opt_param_op: bool = True):
+        """Sets the `opt_param_op` flag.
+
+        Args:
+            opt_param_op (bool): True, if operator has trainable parameters
+        """
+        self._opt_param_op = opt_param_op
+        self._l1.set_opt_param_op(opt_param_op)
+        self._l2.set_opt_param_op(opt_param_op)
+
+    @property
+    def loss_variance_available(self) -> bool:
+        if self._composition in ("*", "/"):
+            return False
+        else:
+            return self._l1.loss_variance_available and self._l2.loss_variance_available
+
+    @property
+    def loss_args_tuple(self) -> tuple:
+        """Returns evaluation tuple for composed loss calculation."""
+        return tuple(set(self._l1.loss_args_tuple + self._l2.loss_args_tuple))
+
+    @property
+    def variance_args_tuple(self) -> tuple:
+        """Returns evaluation tuple for composed variance calculation."""
+        if self._composition in ("*", "/"):
+            raise ValueError("Variance not available for composition: ", self._composition)
+        else:
+            return tuple(set(self._l1.variance_args_tuple + self._l2.variance_args_tuple))
+
+    @property
+    def gradient_args_tuple(self) -> tuple:
+        """Returns evaluation tuple for composed gradient calculation."""
+        return tuple(set(self._l1.gradient_args_tuple + self._l2.gradient_args_tuple))
+
+    def value(self, value_dict: dict, **kwargs) -> float:
+        """Calculates and returns the composed loss value.
+
+        Args:
+            value_dict (dict): Dictionary with values for the evaluation of the loss function
+
+        Returns:
+            float: Composed loss value
+        """
+
+        value_l1 = self._l1.value(value_dict, **kwargs)
+        value_l2 = self._l2.value(value_dict, **kwargs)
+
+        if self._composition == "*":
+            return value_l1 * value_l2
+        elif self._composition == "/":
+            return value_l1 / value_l2
+        elif self._composition == "+":
+            return value_l1 + value_l2
+        elif self._composition == "-":
+            return value_l1 - value_l2
+        else:
+            raise ValueError("Unknown composition: ", self._composition)
+
+    def variance(self, value_dict: dict, **kwargs) -> float:
+        """Calculates and returns the composed variance value.
+
+        Args:
+            value_dict (dict): Dictionary with values for the evaluation of the loss function
+
+        Returns:
+            float: Composed variance value
+        """
+
+        if self._composition in ("*", "/"):
+            raise ValueError("Variance not available for composition: ", self._composition)
+
+        var_l1 = self._l1.variance(value_dict, **kwargs)
+        var_l2 = self._l2.variance(value_dict, **kwargs)
+
+        if self._composition == "+":
+            return var_l1 + var_l2
+        elif self._composition == "-":
+            return var_l1 + var_l2
+        else:
+            raise ValueError("Unknown composition: ", self._composition)
+
+    def gradient(
+        self, value_dict: dict, **kwargs
+    ) -> Union[np.ndarray, tuple[np.ndarray, np.ndarray]]:
+        """Calculates and returns the gradient of the composed loss.
+
+        Args:
+            value_dict (dict): Dictionary with values for the evaluation of the
+                loss function gradient
+
+        Returns:
+            Union[np.ndarray, tuple[np.ndarray, np.ndarray]]: Gradient of the composed
+                loss function
+
+        """
+
+        grad_l1 = self._l1.gradient(value_dict, **kwargs)
+        grad_l2 = self._l2.gradient(value_dict, **kwargs)
+        if self._composition in ("*", "/"):
+            value_l1 = self._l1.value(value_dict, **kwargs)
+            value_l2 = self._l2.value(value_dict, **kwargs)
+
+        if isinstance(grad_l1, tuple) and isinstance(grad_l2, tuple):
+            if self._composition == "*":
+                # (f*g)' = f'*g + f*g'
+                return tuple(
+                    [
+                        np.add(grad_l1[i] * value_l2, grad_l2[i] * value_l1)
+                        for i in range(len(grad_l1))
+                    ]
+                )
+            elif self._composition == "/":
+                # (f/g)' = (f'*g - f*g')/g^2
+                return tuple(
+                    [
+                        np.subtract(grad_l1[i] / value_l2, value_l1 / value_l2 * grad_l2[i])
+                        for i in range(len(grad_l1))
+                    ]
+                )
+            elif self._composition == "+":
+                return tuple([np.add(grad_l1[i], grad_l2[i]) for i in range(len(grad_l1))])
+            elif self._composition == "-":
+                return tuple([np.subtract(grad_l1[i], grad_l2[i]) for i in range(len(grad_l1))])
+            else:
+                raise ValueError("Unknown composition: ", self._composition)
+
+        elif not isinstance(grad_l1, tuple) and not isinstance(grad_l2, tuple):
+            if self._composition == "*":
+                # (f*g)' = f'*g + f*g'
+                return np.add(grad_l1 * value_l2, grad_l2 * value_l1)
+            elif self._composition == "/":
+                # (f/g)' = (f'*g - f*g')/g^2
+                return np.subtract(grad_l1 / value_l2, value_l1 / value_l2 * grad_l2)
+            elif self._composition == "+":
+                return np.add(grad_l1, grad_l2)
+            elif self._composition == "-":
+                return np.subtract(grad_l1, grad_l2)
+            else:
+                raise ValueError("Unknown composition: ", self._composition)
+        else:
+            raise ValueError("Gradient output structure types do not match!")
+
+class ConstantLoss(QNNLossBase):
+    """Class for constant or independent loss functions.
+
+    Args:
+        value (Union[int, float, Callable[[int],float]]): Constant value or function depending
+            on the iterations returning a constant value.
+    """
+
+    def __init__(self, value: Union[int, float, Callable[[int], float]] = 0.0):
+        super().__init__()
+        if callable(value):
+            self._value = value
+        else:
+            self._value = float(value)
+
+    @property
+    def loss_variance_available(self) -> bool:
+        """Returns True if the loss function has a variance function."""
+        return True
+
+    @property
+    def loss_args_tuple(self) -> tuple:
+        """Returns empty evaluation tuple for loss calculation."""
+        return tuple()
+
+    @property
+    def variance_args_tuple(self) -> tuple:
+        """Returns empty evaluation tuple for variance calculation."""
+        return tuple()
+
+    @property
+    def gradient_args_tuple(self) -> tuple:
+        """Returns empty evaluation tuple for gradient calculation."""
+        return tuple()
+
+    def value(self, value_dict: dict, **kwargs) -> float:
+        """Returns constant or iteration dependent loss value
+
+        Args:
+            value_dict (dict): Contains calculated values of the model
+            iteration (int): iteration number, if value is a callable function
+        """
+        if callable(self._value):
+            if "iteration" not in kwargs:
+                raise AttributeError("If value is callable, iteration is required.")
+            return self._value(kwargs["iteration"])
+        return self._value
+
+    def variance(self, value_dict: dict, **kwargs) -> float:
+        """Returns zero variance of the constant loss function."""
+        return 0.0
+
+    def gradient(
+        self, value_dict: dict, **kwargs
+    ) -> Union[np.ndarray, tuple[np.ndarray, np.ndarray]]:
+        """Returns zero gradient value
+
+        Args:
+            value_dict (dict): Contains calculated values of the model
+        """
+        dp = np.zeros(value_dict["param"].shape)
+        dop = np.zeros(value_dict["param_op"].shape)
+        if self._opt_param_op:
+            return dp, dop
+        return dp
