+++ src/squlearn/qnn/qnnr.py
@@ -9,7 +9,8 @@ from sklearn.base import RegressorMixin
 from tqdm import tqdm
 
 from .base_qnn import BaseQNN
-from .loss import LossBase, VarianceLoss
+from .loss import VarianceLoss
+from .loss.qnn_loss_base import QNNLossBase
 from .training import train_mini_batch, train, ShotControlBase
 
 from ..observables.observable_base import ObservableBase
@@ -34,7 +35,7 @@ class QNNRegressor(BaseQNN, RegressorMixin):
             is used in the expectation value of the QNN. Can be a list for multiple outputs. For a
             list of operators, check this list of implemented :ref:`operators`.
         executor (Executor): Executor instance.
-        loss (LossBase): The loss function to be optimized. Can also be combination of multiple
+        loss (QNNLossBase): The loss function to be optimized. Can also be combination of multiple
             loss functions.
         optimizer (OptimizerBase): The optimizer instance that is used to minimize the loss
             function.
@@ -155,7 +156,7 @@ class QNNRegressor(BaseQNN, RegressorMixin):
         encoding_circuit: EncodingCircuitBase,
         operator: Union[ObservableBase, list[ObservableBase]],
         executor: Executor,
-        loss: LossBase,
+        loss: QNNLossBase,
         optimizer: OptimizerBase,
         param_ini: np.ndarray = None,
         param_op_ini: np.ndarray = None,
